Add specs for marriages
