+++ Gemfile
@@ -38,7 +38,7 @@ gem 'attr_encrypted', '3.1.0'
 gem 'aws-sdk-s3', '~> 1'
 gem 'aws-sdk-sns', '~> 1'
 gem 'betamocks', git: 'https://github.com/department-of-veterans-affairs/betamocks', branch: 'master'
-gem 'bgs_ext', git: 'https://github.com/department-of-veterans-affairs/bgs-ext.git', require: 'bgs'
+gem 'bgs_ext', git: 'https://github.com/department-of-veterans-affairs/bgs-ext.git', require: 'bgs', path: '/Users/derekdyer/Documents/bgs-ext', branch: 'adjust_vnp_benefit_claim_proc_id'
 gem 'breakers'
 gem 'carrierwave-aws'
 gem 'clam_scan'
