+++ config/settings/test.yml
@@ -73,7 +73,10 @@ bgs:
   application: VAgovAPI
   client_station_id: 281
   client_username: VAgovAPI
-  url: https://internal-dsva-vagov-dev-fwdproxy-1893365470.us-gov-west-1.elb.amazonaws.com:4447
+  ssl_verify_mode: 'none'
+  log: true
+  url: https://localhost:4447
+#  url: https://internal-dsva-vagov-dev-fwdproxy-1893365470.us-gov-west-1.elb.amazonaws.com:4447
 
 flipper:
   admin_user_emails:
