+++ spec/lib/bgs/dependents_spec.rb
@@ -92,82 +92,6 @@ RSpec.describe BGS::Dependents do
       end
     end
 
-    context 'adding a spouse' do
-      it 'returns hash for spouse who lives with veteran' do
-        json = File.read("#{fixtures_path}/spouse/spouse_lives_with_veteran.json")
-        payload = JSON.parse(json)
-
-        VCR.use_cassette('bgs/dependents/create/spouse/lives_with_veteran') do
-          dependents = BGS::Dependents.new(
-            proc_id: proc_id,
-            payload: payload,
-            user: user_hash
-          ).create
-
-          expect(dependents).to include(
-            a_hash_including(
-              family_relationship_type_name: 'Spouse',
-              participant_relationship_type_name: 'Spouse',
-              marriage_city: 'Slawson',
-              marriage_state: 'CA',
-              type: 'spouse',
-              begin_date: '2014-03-04'
-            )
-          )
-        end
-      end
-
-      it 'returns hash for spouse who has different address (separated)' do
-        VCR.use_cassette('bgs/dependents/create') do
-          dependents = BGS::Dependents.new(
-            proc_id: proc_id,
-            payload: all_flows_payload,
-            user: user_hash
-          ).create
-
-          expect(dependents).to include(
-            a_hash_including(
-              family_relationship_type_name: 'Estranged Spouse',
-              participant_relationship_type_name: 'Spouse',
-              marriage_city: 'Slawson',
-              marriage_state: 'CA',
-              type: 'spouse',
-              begin_date: '2014-03-04'
-            )
-          )
-        end
-      end
-
-      it 'marks spouse as veteran' do
-        json = File.read("#{fixtures_path}/spouse/spouse_is_veteran.json")
-        payload = JSON.parse(json)
-        spouse_vet_hash = {
-          'first' => 'Jenny',
-          'middle' => 'Lauren',
-          'last' => 'McCarthy',
-          'suffix' => 'Sr.',
-          'ssn' => '323454323',
-          'birth_date' => '1981-04-04',
-          'ever_married_ind' => 'Y',
-          'vet_ind' => 'Y',
-          'va_file_number' => '00000000',
-          'martl_status_type_cd' => 'Married'
-        }
-
-        VCR.use_cassette('bgs/dependents/create/spouse/is_veteran') do
-          expect_any_instance_of(BGS::Service).to receive(:create_person)
-            .with('3828033', '149487', spouse_vet_hash)
-            .and_call_original
-
-          BGS::Dependents.new(
-            proc_id: proc_id,
-            payload: payload,
-            user: user_hash
-          ).create
-        end
-      end
-    end
-
     xcontext 'reporting a divorce' do
       it 'returns an hash with divorce data' do
         VCR.use_cassette('bgs/dependents/create') do
